feat: extract baserowtable into trait
