+++ examples/table_by_id.rs
@@ -1,6 +1,9 @@
 use std::env;
 
-use baserow_rs::{api::client::BaserowClient, filter::Filter, Baserow, ConfigBuilder, OrderDirection};
+use baserow_rs::{
+    api::client::BaserowClient, filter::Filter, Baserow, BaserowTableOperations, ConfigBuilder,
+    OrderDirection,
+};
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
