+++ src/api/mod.rs
@@ -2,3 +2,4 @@ pub mod authentication;
 pub mod client;
 pub mod file;
 pub mod table;
+pub mod table_operations;
